requirements.txt
langchain 
langchain-core
langchain-community
pypdf
pymupdf
sentence-transformers
faiss-cpu
chromadb
google.generativeai
langchain_chroma